requirement.txt
Django>=5.1,<5.2
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-filter
django-extensions
psycopg2-binary
django-celery-beat
redis